Added a Curse Basic Functionality
Added a pillar prefab that creates an area which gives a player a curse. The curse can reduce health by 10% per stack, where stack is a count of triggering the curse source (by colliding with its area). Curse are stored in a list of a player and will be deleted when player uses a purifier - a special child object of a Curse Manager. Also, added a text for displaying player's health and maxHealth. Also, a Curse is a dataType that contains its type (Health or Damage) and a stacks count.