Added a new button texture with selection possibility
Added 3 button images and a script that allows you to select and deselect buttons using W and S keys, Space to click. Worked with a resolution 1920x1080 so the project may contain UI bugs such as big buttons. Added a company name, changed scene's name to "Village". Fixed a bug where MainManager instance was not found.